Add Array component
The Array component implements an HTML5 canvas object to display a
given array of numbers in the range [0, 1) as vertical or horizontal
bars.

It also takes two more arguments; check and compare indicies. These
indicies indicate where a given sorting algorithm might be acting at a
given moment and the display colour of the bar (at the specified index)
willl change accordingly.