Begin rewrite of code
The previous version was running on reactJS, and as a result had a lot
of (potentially unnecessary) overhead.

The site is very simple, so does not need much more than some native
javascript on a barebones HTML page.

This rewrite will attempt to address these problems associated with the
react version.