Install react redux
React redux is a module for using redux state with react UI. Redux is
not speciffically related to react, and consequently the way we connect
components to state must be implemented by us. React redux handles a lot
of this for us and also makes some "difficult to write by hand"
optimisations to the data flow as well.