Move state off main thread
We have moved the store to a worker thread and just use comlink's expose
and wrap methods to abstract the use of postMessage.

As simple as that!!?? Not quite. Since anything including postMessage
will be inherently async, and redux expects the store to behave
synchronously, the app will not work in its current state.

We have to make some more simple changes, but we have made a start.