Update babel config
Babel will not play nicely with async/await or generator functions. By
setting the target for the env preset, babel will know how we want it to
handle these special parts of our syntax.

Since we will be doing work with workers, and workers are fundamentally
async, we want to use the async/await in our app.