Write a tracker class
The tracker class is intended to be passed as an argument to the
algorithms, for which it will record the moves made on the data.

Sorting algorithms local variables will likely be held in very fast
memory in practice, so any time spent on checks associated with loops
etc. will be negligible. Time will mostly be spent on fetching the
values in the array from memory, which will likely be slower, as they
must persist after the sort completes. In other words, it may be assumed
during analysis that the inner workings of the sort happen infinitely
fast.