Change tracking pattern
Previously we have a Tracker class that was passed to the sorting
algorithm and we *trusted* that the algorithm would not access the
_values object on the tracker. Now we have a calculate method, that
takes advantage of the js closure to only pass the intended apis to the
sorting algorithm.

This change has ALSO changed the pattern for calculating the moves for a
sort. Previously we would create the tracker object and mutate it with
the sorting algorithm. Now we pass both the algorithm and the values to
the calculation method, and it will return the sort moves **without**
mutating either the sorting algorithm *or* the values passed.