Untrack instead of giving moves
The moves array can become very large; if an O(n^2) sort isperformed on
an array of length 2000 for example, we could expect the number of moves
to be of order ~10^6. An array this large should not be copied, and
herein lies the main problem.

Originally we would just return the reference to the moves array. This
is not great because potentially sorting algorithms would be able to
modify the moves that they had been performing. For example, we could
make a sorting algorithm that makes O(n^2) compares to work out the
order of the data, removes all of these compare moves from the moves
array and then performs O(n) swaps to sort the data correctly. This is
cheating and it is not good.

We now return an Untracker which wraps the moves and provides simple
apis for navigating the data collected.