Prevent default key event
If there is a button which should be activated by a key, then the key
should not activate anything else.

For example, suppose a user used the mouse to activate play/pause. The
button for play/pause would not be in focus. If the user then pressed
space, the button would be pressed once because it is assigned to space
and again because it is in focus; nothing would happen.