Update memcpy
Each move should be O(1), and memcpy of a chunk is O(n) where n is the
length of the chunk. Therefore we should only move chunks of length 1 in
a single move.

Also, we must remember the original value that was overwritten so that
we can watch the sort in reverse.