Implement transferable moves
Calculating sorts off the main thread is all well and good, but the main
thread still blocks if we use structured cloning to transport the
calculated moves to the main thread.

By implementing the moves in an ArrayBuffer, we can transfer the moves
from the worker to the main thread instead of copying them. This vastly
improves the user experience as the main thread will NOT block any more.