Re-implement sort changing
Previously, when a new sort was chosen, it would calculate on the
current state of the values. This meant resetting the unsorted array to
the current array (inadvertently causing the previously selected sort to
be calculated again on the current data) and then changing the sort to
the new algorithm before calculating the new sort.

Now, instead of the new sort being calculated on the current values, it
is calculated on the same original data as the current sort. The curent
sort is rewound and then the new sort is played after it.