Cancel empty reset
If a sort is reset, but has not made any moves yet, the screen will
flash grey for a moment while the user action block is obtained, the
unchanged values array is reset (doing nothing) and the block is
released. This appears as a single sharp UI flash, not pleasing to the
eye. By cancelling the reset on unchanged data, this problem is avoided.